refactor(android): split v2 shell screens
